working

+++ utils/metrics.py
@@ -42,8 +42,8 @@ def fast_confusion(true, pred, label_values=None):
     """
 
     # Ensure data is in the right format
-    true = np.squeeze(true)
-    pred = np.squeeze(pred)
+    true = np.squeeze(true).astype(np.int32)
+    pred = np.squeeze(pred).astype(np.int32)
     if len(true.shape) != 1:
         raise ValueError('Truth values are stored in a {:d}D array instead of 1D array'. format(len(true.shape)))
     if len(pred.shape) != 1:
